fix: telemetry fields that failed whole new-gen report frames
New-generation pushalls send canvas_id/batch_id as integers and nozzle
wear as a float; either failed TelemetryReport deserialization, so every
A2L/H2/P2S/X2D report surfaced as TelemetryEvent::Unknown. The cloud-id
strings now use the permissive number-or-string deserializer and wear is
f32.

stat is 9-13 hex digits on those models; the u32 parse overflowed and
reported every door as closed. It now parses as u64.

AmsUnit temp/humidity are optional and preserved on absence, as
BambuStudio's ParseAmsInfo treats them, so a partial unit push no longer
drops the frame.

Closes #339, Closes #340, Closes #341, Closes #338

+++ src/types/telemetry/device.rs
@@ -319,7 +319,10 @@ pub struct NozzleInfo {
     pub nozzle_type: Option<String>,
 
     /// Normalized physical wear tracker value.
-    pub wear: Option<u32>,
+    ///
+    /// A float: H2C, P2S and X2D send `0.0`, and BambuStudio stores it as `float m_wear`
+    /// (`DevNozzleSystem.h:104`).
+    pub wear: Option<f32>,
 
     /// Hotend manufacturer serial number (verbose IDEX platform representation).
     pub serial_number: Option<String>,
